# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Sensei MVP API",
    "description": "API for the Sensei Product Sense Practice Game - help users practice structured product thinking with AI-powered feedback",
    "version": "1.0.0",
    "contact": {
      "name": "Sensei Team"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
...
  "paths": {
    "/auth/send-link": {
      "post": {
        "tags": ["Authentication"],
        "summary": "Send magic link to user email",
        "description": "Generates a magic link token and sends it to the user's email (or logs to console in dev mode)",
        "operationId": "sendMagicLink",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email"],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email",
                    "example": "user@example.com"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Magic link sent successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                },
                "example": {
                  "data": {
                    "message": "Magic link sent to user@example.com",
                    "expiresIn": "15m"
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
    "/account": {
      "delete": {
        "tags": ["Account"],
        "summary": "Delete user account",
        "description": "Permanently deletes user account and all associated data (sessions, XP, etc.)",
        "operationId": "deleteAccount",
        "security": [{ "BearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["confirmation"],
                "properties": {
                  "confirmation": {
                    "type": "string",
                    "enum": ["DELETE"],
                    "description": "Must type 'DELETE' to confirm"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Account deleted successfully"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /auth/verify endpoint"
      }
    },
    "schemas": {
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object"
          },
          "meta": {
            "type": "object",
            "properties": {
              "timestamp": {
                "type": "string",
                "format": "date-time"
              },
              "requestId": {
                "type": "string",
                "format": "uuid"
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "VALIDATION_ERROR"
              },
              "message": {
                "type": "string",
                "example": "Invalid input provided"
              },
              "details": {
                "type": "object"
              }
            },
            "required": ["code", "message"]
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "VALIDATION_ERROR",
                "message": "Invalid input provided",
                "details": {
                  "fields": {
                    "email": "Invalid email format"
                  }
                }
              }
            }
          }
        }
      },
      "UnauthorizedError": {
        "description": "Unauthorized - invalid or missing token",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "UNAUTHORIZED",
                "message": "Invalid or expired token"
              }
            }
          }
        }
      },
...
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "error": {
                "code": "INTERNAL_ERROR",
                "message": "An unexpected error occurred"
              }
            }
          }
        }
      }
    }
  }
}